Added custom-exception with resolved issues

+++ src/main/java/com/stackroute/springboot/muzix/service/TrackServiceImpl.java
@@ -1,5 +1,7 @@
 package com.stackroute.springboot.muzix.service;
 
+import com.stackroute.springboot.muzix.exception.TrackAlreadyExistsException;
+import com.stackroute.springboot.muzix.exception.TrackNotFoundException;
 import com.stackroute.springboot.muzix.model.Track;
 import com.stackroute.springboot.muzix.repository.TrackRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,32 +25,47 @@ public class TrackServiceImpl implements TrackService {
 
     //method for save the track
     @Override
-    public Track saveTrack(Track track) {
+    public Track saveTrack(Track track) throws TrackAlreadyExistsException {
 
-        Track saveTrack = trackRepository.save(track);
+        Track saveTrack = null;
+
+        if(trackRepository.existsById(track.getTrackId()))
+            throw new TrackAlreadyExistsException("Track Already exists");
+        else {
+            saveTrack = trackRepository.save(track);
+        }
+
+
         return saveTrack;
     }
 
     //method for get all the tracks
     @Override
-    public List<Track> getAllTracks() {
+    public List<Track> getAllTracks() throws TrackNotFoundException {
         return trackRepository.findAll();
     }
 
     //method for update a new track
     @Override
-    public Track updateTrack(Track track){
+    public Track updateTrack(Track track) throws TrackNotFoundException{
 
-     /*  Integer  id = track.getTrackId();
-        Optional<Track> userOptional = trackRepository.findById(id);*/
-        trackRepository.save(track);
+        if(!trackRepository.existsById(track.getTrackId())) {
+            throw new TrackNotFoundException("Track Not Found Exception");
+        }
+        else {
+            trackRepository.save(track);
+        }
         return track;
 
     }
     @Override
-    public Track deleteTrack(Track track) {
-
-        trackRepository.deleteById(track.getTrackId());
+    public Track deleteTrack(Track track) throws TrackNotFoundException{
+        if(!trackRepository.existsById(track.getTrackId())) {
+            throw new TrackNotFoundException("Track Not Found Exception");
+        }
+        else{
+            trackRepository.deleteById(track.getTrackId());
+        }
         return track;
     }
     @Override
@@ -57,7 +74,7 @@ public class TrackServiceImpl implements TrackService {
     }
 
     @Override
-    public List<Track> trackByName(String name) {
+    public List<Track> trackByName(String name){
         List<Track> tracks=trackRepository.findTitleByName(name);
         return tracks;
     }
